# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Prueba Técnica API",
        "description": "Documentación de la API para la prueba técnica.",
        "contact": {
            "name": "Tu Nombre",
            "email": "tuemail@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/GetAllTasks": {
            "get": {
                "summary": "List all tasks",
                "operationId": "94554ec613042af56cd51b2ecca0a6f2",
                "responses": {
                    "200": {
                        "description": "A list of tasks",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Task"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/SaveTasks": {
            "post": {
                "summary": "Create a new task",
                "operationId": "b6660a5723c750f8e6d8103cc02937e7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Task"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Task created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Task"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/GetTasks/{id}": {
            "get": {
                "summary": "Get a task by ID",
                "operationId": "a47a1160975182e79b065cc57ae27475",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the task to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A task",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Task"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Task not found"
                    }
                }
            }
        },
        "/UpdateTasks/{id}": {
            "put": {
                "summary": "Update a task by ID",
                "operationId": "687a1b5c4f40a9b4729db2171822267b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the task to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Task"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Task updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Task"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Task not found"
                    }
                }
            }
        },
        "/DeleteTasks/{id}": {
            "delete": {
                "summary": "Delete a task by ID",
                "operationId": "91f7adf3953f4ec60d25b582c4a1aab3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the task to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Task deleted successfully"
                    },
                    "404": {
                        "description": "Task not found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Task": {
                "required": [
                    "id",
                    "title",
                    "description",
                    "due_date"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Task Title"
                    },
                    "description": {
                        "type": "string",
                        "example": "Task Description"
                    },
                    "due_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2024-12-31"
                    }
                },
                "type": "object"
            }
        }
    }
}